Allowing Valid Callbacks to get fired in IE9
When using a valid callback and saying return false browsers that did
not support checkValidity (ie) just hung and did nothing. Now I've
changed the logic so that if it doesn't support checkValidity it calls
options.valid if the form is submitted. Not the perfect solution but
fixes many cart related problems for us in IE.